+++ contracts/src/v0.8/interfaces/VRFCoordinatorV2Interface.sol
@@ -3,32 +3,60 @@ pragma solidity ^0.8.0;
 
 interface VRFCoordinatorV2Interface {
 
-    function requestRandomWords(
-        bytes32 keyHash,  // Corresponds to a particular offchain job which uses that key for the proofs
-        uint64  subId,   // A data structure for billing
-        uint16  minimumRequestConfirmations,
-        uint32  callbackGasLimit,
-        uint32  numWords,  // Desired number of random words
-        uint32 consumerID
-    )
+  function requestRandomWords(
+    bytes32 keyHash,  // Corresponds to a particular offchain job which uses that key for the proofs
+    uint64  subId,   // A data structure for billing
+    uint16  minimumRequestConfirmations,
+    uint32  callbackGasLimit,
+    uint32  numWords,  // Desired number of random words
+    uint32 consumerID
+  )
     external
-    returns (uint256 requestId);
+    returns (
+      uint256 requestId
+    );
 
-    function createSubscription(
-        address[] memory consumers // permitted consumers of the subscription
-    )
+  function createSubscription(
+    address[] memory consumers // permitted consumers of the subscription
+  )
     external
-    returns (uint64 subId);
+    returns (
+      uint64 subId
+    );
 
-    function fundSubscription(
-        uint64 subId,
-        uint96 amount
-    )
+  function getSubscription(
+    uint64 subId
+  )
+    external
+    view
+    returns (
+      uint96 balance,
+      address owner,
+      address[] memory consumers
+    );
+
+  function fundSubscription(
+    uint64 subId,
+    uint96 amount
+  )
     external;
 
-    function updateSubscription(
-        uint64 subId,
-        address[] memory consumers
-    )
+  function updateSubscription(
+    uint64 subId,
+    address[] memory consumers
+  )
+    external;
+
+  function defundSubscription(
+    uint64 subId,
+    address to,
+    uint96 amount
+  )
+    external;
+
+  function cancelSubscription(
+    uint64 subId,
+    address to
+  )
     external;
 }